Add bootstrap list for articles to have better test data Add row number column
