+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/entity/ArticlesController.java
@@ -5,32 +5,20 @@ import de.digitalcollections.cudami.admin.util.LanguageSortingHelper;
 import de.digitalcollections.cudami.client.CudamiClient;
 import de.digitalcollections.cudami.client.CudamiLocalesClient;
 import de.digitalcollections.cudami.client.identifiable.entity.CudamiArticlesClient;
-import de.digitalcollections.model.exception.ResourceNotFoundException;
 import de.digitalcollections.model.exception.TechnicalException;
 import de.digitalcollections.model.identifiable.entity.Article;
-import de.digitalcollections.model.identifiable.resource.FileResource;
-import de.digitalcollections.model.list.paging.PageRequest;
-import de.digitalcollections.model.list.paging.PageResponse;
-import de.digitalcollections.model.list.sorting.Order;
-import de.digitalcollections.model.list.sorting.Sorting;
 import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 /** Controller for articles management pages. */
 @Controller
@@ -52,12 +40,6 @@ public class ArticlesController extends AbstractController {
   public String create(Model model) throws TechnicalException {
     model.addAttribute("activeLanguage", localeService.getDefaultLanguage());
     return "articles/create";
-  }
-
-  @GetMapping("/api/articles/new")
-  @ResponseBody
-  public Article create() throws TechnicalException {
-    return service.create();
   }
 
   @GetMapping("/articles/{uuid}/edit")
@@ -82,28 +64,6 @@ public class ArticlesController extends AbstractController {
     return "articles/edit";
   }
 
-  @GetMapping("/api/articles")
-  @ResponseBody
-  public PageResponse<Article> find(
-      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
-      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
-      @RequestParam(name = "searchTerm", required = false) String searchTerm,
-      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
-      throws TechnicalException {
-    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
-    if (sortBy != null) {
-      Sorting sorting = new Sorting(sortBy);
-      pageRequest.setSorting(sorting);
-    }
-    return this.service.find(pageRequest);
-  }
-
-  @GetMapping("/api/articles/{uuid}")
-  @ResponseBody
-  public Article getByUuid(@PathVariable UUID uuid) throws TechnicalException {
-    return service.getByUuid(uuid);
-  }
-
   @GetMapping("/articles")
   public String list(Model model) throws TechnicalException {
     final Locale locale = LocaleContextHolder.getLocale();
@@ -116,45 +76,5 @@ public class ArticlesController extends AbstractController {
   @ModelAttribute("menu")
   protected String module() {
     return "articles";
-  }
-
-  @PostMapping("/api/articles")
-  public ResponseEntity save(@RequestBody Article article) {
-    try {
-      Article articleDb = service.save(article);
-      return ResponseEntity.status(HttpStatus.CREATED).body(articleDb);
-    } catch (TechnicalException e) {
-      LOGGER.error("Cannot save article: ", e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-    }
-  }
-
-  @PutMapping("/api/articles/{uuid}")
-  public ResponseEntity update(@PathVariable UUID uuid, @RequestBody Article article) {
-    try {
-      Article articleDb = service.update(uuid, article);
-      return ResponseEntity.ok(articleDb);
-    } catch (TechnicalException e) {
-      LOGGER.error("Cannot save article with uuid={}", uuid, e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-    }
-  }
-
-  @GetMapping("/articles/{uuid}")
-  public String view(@PathVariable UUID uuid, Model model)
-      throws TechnicalException, ResourceNotFoundException {
-    final Locale displayLocale = LocaleContextHolder.getLocale();
-    Article article = service.getByUuid(uuid);
-    if (article == null) {
-      throw new ResourceNotFoundException();
-    }
-    List<Locale> existingLanguages =
-        languageSortingHelper.sortLanguages(displayLocale, article.getLabel().getLocales());
-    List<FileResource> relatedFileResources = service.getRelatedFileResources(article.getUuid());
-    model
-        .addAttribute("article", article)
-        .addAttribute("existingLanguages", existingLanguages)
-        .addAttribute("relatedFileResources", relatedFileResources);
-    return "articles/view";
   }
 }
