+++ dc-cudami-admin/src/main/resources/messages.properties
@@ -105,9 +105,11 @@ last_modified=last modified
 lastname=Lastname
 lbl.actions=Actions
 lbl.created=created
+lbl.label=Label
 lbl.last_modified=last modified
 lbl.predicate.inline=relation type
-lbl.uuid=ID
+lbl.row_number=#
+lbl.uuid=UUID
 lbl.value=Value
 lbl_password=Password
 lbl_stay_logged_in=Stay logged in
