+++ dc-cudami-admin/src/main/resources/messages_de.properties
@@ -105,9 +105,11 @@ last_modified=letzte Änderung
 lastname=Nachname
 lbl.actions=Aktionen
 lbl.created=erstellt
+lbl.label=Bezeichnung
 lbl.last_modified=letzte Änderung
 lbl.predicate.inline=Beziehungstyp
-lbl.uuid=ID
+lbl.row_number=#
+lbl.uuid=UUID
 lbl.value=Wert
 lbl_password=Passwort
 lbl_stay_logged_in=Dauerhaft einloggen
